flytt småbarnstillegg-logikk ut fra BeregningService og inn i egen SmåbarnstilleggService
