+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/BeregningService.kt
@@ -211,42 +211,6 @@ class BeregningService(
         return oppdaterBehandlingMedBeregning(behandling, personopplysningGrunnlag)
     }
 
-    fun kanAutomatiskIverksetteSmåbarnstilleggEndring(
-        behandling: Behandling,
-        sistIverksatteBehandling: Behandling?,
-    ): Boolean {
-        if (!behandling.skalBehandlesAutomatisk || !behandling.erSmåbarnstillegg()) return false
-
-        val forrigeSmåbarnstilleggAndeler =
-            if (sistIverksatteBehandling == null) {
-                emptyList()
-            } else {
-                hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(
-                    behandlingId = sistIverksatteBehandling.id,
-                ).filter { it.erSmåbarnstillegg() }
-            }
-
-        val nyeSmåbarnstilleggAndeler =
-            if (sistIverksatteBehandling == null) {
-                emptyList()
-            } else {
-                hentAndelerTilkjentYtelseMedUtbetalingerForBehandling(
-                    behandlingId = behandling.id,
-                ).filter { it.erSmåbarnstillegg() }
-            }
-
-        val (innvilgedeMånedPerioder, reduserteMånedPerioder) =
-            hentInnvilgedeOgReduserteAndelerSmåbarnstillegg(
-                forrigeSmåbarnstilleggAndeler = forrigeSmåbarnstilleggAndeler,
-                nyeSmåbarnstilleggAndeler = nyeSmåbarnstilleggAndeler,
-            )
-
-        return kanAutomatiskIverksetteSmåbarnstillegg(
-            innvilgedeMånedPerioder = innvilgedeMånedPerioder,
-            reduserteMånedPerioder = reduserteMånedPerioder,
-        )
-    }
-
     /**
      * Henter alle barn på behandlingen som har minst en periode med tilkjentytelse.
      */
